backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.11.0
PyJWT[crypto]==2.10.1
pydantic-settings==2.7.1
httpx==0.28.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
python-multipart==0.0.20
slowapi==0.1.9
PyYAML==6.0.2
google-cloud-aiplatform>=1.60.0
python-json-logger>=4.0.0
Pillow>=10.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
